// examin_rest_api_plus_frontend/examin-web-front-master/src/app/models/user.model.ts
export class User {
    userId?: number;
    instituteId?: number;
    username?: string;
    mobile?: string;
    fullName?: string;
    role?: string;
    dateOfBirth?: string;
    profileImage?: string;
    address?: object | boolean;
    profileCompleted?: boolean;
}

// TODO: cleanup model if not in use